requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
Django==3.0.3
django-braces==1.14.0
django-filter==2.3.0
django-phonenumber-field==4.0.0
phonenumbers==8.12.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
